+++ food-retail-marketplace/ardo-backend/service/management/management_product_service.go
@@ -11,8 +11,8 @@ import (
 type ManagementProductService interface {
 	GetProductsByFilters(ctx context.Context, offset, limit int64, isDeleted bool) ([]*product.Product, int64, error)
 	GetProductById(ctx context.Context, productId string) (*product.Product, error)
-	AddProduct(ctx context.Context, name, desc core.MlString, price float64, quantity int, img, status string) error
-	UpdateProduct(ctx context.Context, productId string, name, desc core.MlString, price float64, quantity int, img, status string) error
+	AddProduct(ctx context.Context, name, desc core.MlString, price float64, quantity int64, img, status string) error
+	UpdateProduct(ctx context.Context, productId string, name, desc core.MlString, price float64, quantity int64, img, status string) error
 	DeleteProduct(ctx context.Context, productId string) error
 	RecoverProduct(ctx context.Context, productId string) error
 }
@@ -38,7 +38,7 @@ func (m *managementProductService) GetProductById(ctx context.Context, productId
 	return m.productService.GetById(ctx, productId)
 }
 
-func (m *managementProductService) AddProduct(ctx context.Context, name, desc core.MlString, price float64, quantity int, img, status string) error {
+func (m *managementProductService) AddProduct(ctx context.Context, name, desc core.MlString, price float64, quantity int64, img, status string) error {
 	newProduct, err := product.NewProduct(name, desc, price, quantity, img, status)
 	if err != nil {
 		return err
@@ -46,7 +46,7 @@ func (m *managementProductService) AddProduct(ctx context.Context, name, desc co
 	return m.productService.Create(ctx, newProduct)
 }
 
-func (m *managementProductService) UpdateProduct(ctx context.Context, productId string, name, desc core.MlString, price float64, quantity int, img, status string) error {
+func (m *managementProductService) UpdateProduct(ctx context.Context, productId string, name, desc core.MlString, price float64, quantity int64, img, status string) error {
 	foundProduct, err := m.productService.GetById(ctx, productId)
 	if err != nil {
 		return err
